Dev: Refactored code.
